Fixs build

+++ travis.sh
@@ -27,7 +27,13 @@ cp -rf include/*            addons/sourcemod/scripting/include
 cp -rf minigames            addons/sourcemod/scripting
 cp -rf MiniGames.sp         addons/sourcemod/scripting
 
-addons/sourcemod/scripting/spcomp -E -v0 addons/sourcemod/scripting/MiniGames.sp        -o"build/plugins/MiniGames.smx"
+addons/sourcemod/scripting/spcomp -E -v0 addons/sourcemod/scripting/MiniGames.sp -o"build/plugins/MiniGames.smx"
+
+if [ ! -f "build/plugins/MiniGames.smx" ]; then
+    echo "Compile failed!"
+    exit 1;
+fi
+
 
 mv LICENSE build
 mv include              build/scripting
